pipewatch-server-develop/pyproject.toml
[build-system]
requires = ["poetry-core>=1.6"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "pipewatch-server"
authors = ["Manuel Schmuck <manuel.schmuck@axelera.ai>"]
description = ""
version = "0.3.8"
readme = "README.md"
packages = [{ include = "pipewatch_server" }]

[tool.poetry.dependencies]
python = "^3.10"
package-lib = { version = "^0.2.4", extras = ["api", "config", "utils"] }
pipewatch-lib = "^0.3.0"
rich = "^13.5.2"
pydantic = "^2.5.0"
pydantic-settings = "^2.0.3"
fastapi = "^0.104.0"
toml = "^0.10.2"
pyyaml = "^6.0.1"
uvicorn = "^0.24.0"
python-gitlab = "^4.1.1"
jinja2 =  "^3.1.3"
mysql-connector-python = "^8.2.0"
